Created new file for class Constants
