+++ todo_tig169/lib/Constants.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class Constants {
+  static const String All = "all";
+  static const String Done = "done";
+  static const String UnDone = "undone";
+
+  static const List<String> choices = <String>[All, Done, UnDone];
+} //constants for the popupmenu
